requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# manage.py: 8
# search/admin.py: 1
# search/apps.py: 3
# search/models.py: 3
# search/tests.py: 1
# search/urls.py: 1
# search/views.py: 2
# whois/urls.py: 16
# whois/wsgi.py: 12
Django == 1.9.7

# search/views.py: 3,4
djangorestframework == 3.4.0

# whois/urls.py: 17
edx-search == 1.0.0
